[#177] Refactor WorkOrderStatusEnum
